refactors folder structure for pr #721
