+++ app/cms/components/Preview.jsx
@@ -1,8 +1,6 @@
 import React, {Component} from "react";
 import {Callout, Icon} from "@blueprintjs/core";
 
-import "./Preview.css";
-
 const previewHash = {
   geo: ["04000US25", "01000US", "04000US36", "05000US25025", "31000US14460", "16000US0455000"],
   cip: ["110701"],
