+++ app/cms/components/cards/GeneratorCard.css
