+++ app/cms/components/editors/GeneratorEditor.css
