+++ app/cms/components/editors/GeneratorEditor.jsx
