+++ app/cms/profile/ProfileEditor.jsx
@@ -2,12 +2,12 @@ import axios from "axios";
 import React, {Component} from "react";
 import {Button, Card, Icon, NonIdealState, Spinner} from "@blueprintjs/core";
 import PropTypes from "prop-types";
-import Loading from "components/Loading";
+import Loading from "../../components/Loading";
 
-import GeneratorCard from "./components/GeneratorCard";
-import TextCard from "./components/TextCard";
-import VisualizationCard from "./components/VisualizationCard";
-import MoveButtons from "./components/MoveButtons";
+import GeneratorCard from "../components/cards/GeneratorCard";
+import TextCard from "../components/cards/TextCard";
+import VisualizationCard from "../components/cards/VisualizationCard";
+import MoveButtons from "../components/MoveButtons";
 
 import "./ProfileEditor.css";
