+++ app/cms/story/StoryTopicEditor.jsx
@@ -1,12 +1,10 @@
 import axios from "axios";
 import React, {Component} from "react";
 import {Button} from "@blueprintjs/core";
-import Loading from "components/Loading";
-import TextCard from "./components/TextCard";
-import VisualizationCard from "./components/VisualizationCard";
-import MoveButtons from "./components/MoveButtons";
-
-import "./StoryTopicEditor.css";
+import Loading from "../../components/Loading";
+import TextCard from "../components/cards/TextCard";
+import VisualizationCard from "../components/cards/VisualizationCard";
+import MoveButtons from "../components/MoveButtons";
 
 const propMap = {
   storytopic_stat: "stats",
